Migrate Accumulo tests to use BaseConnectorTest
Summary of the changes:
- TestAccumuloConnectorTest.java has been renamed
from TestAccumuloDistributedQueries.java
- the content of TestAccumuloIntegrationSmokeTest.java has been
moved inside TestAccumuloConnectorTest.java
